python-sdk/requirements.txt
amay-key-vault-sdk>=1.0.0
requests>=2.25.0 